Not sure how this type error got missed since the IDE pointed it out.

+++ still_cleaner.py
@@ -33,7 +33,7 @@ def main() -> None:
     # If we had 24 hrs of daylight there'd be this many in the continuous folder per day
     # Since we don't live in extreme latitudes that's not the case and this is a safe
     # upper limit that'll let us get through any backlog, eventually.
-    batch_size = (24 * 60) / 5
+    batch_size = int((24 * 60) / 5)
 
     cameras = ["backdoor", "backpatio", "continuous/backdoor", "continuous/backpatio"]
